pyproject.toml
[tool.poetry]
name = "vivpro-assignment"
version = "0.1.0"
description = "A take home assignment"
authors = ["AnkushGochke <ankushgochke@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
djangorestframework = "^3.14.0"
Django = "^5.0.1"


[tool.poetry.group.dev.dependencies]
ipython = "^8.20.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
